Extracted MetricsDeriver class - (Task #346)
---
Task #346: Extract MetricsDeriver class
