+++ de.dlr.sc.virsat.fdir.core/src/de/dlr/sc/virsat/fdir/core/metrics/SteadyStateAvailability.java
@@ -9,10 +9,7 @@
  *******************************************************************************/
 package de.dlr.sc.virsat.fdir.core.metrics;
 
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 
 import org.eclipse.core.runtime.SubMonitor;
 
@@ -23,14 +20,13 @@ import de.dlr.sc.virsat.fdir.core.matrix.iterator.IMatrixIterator;
 import de.dlr.sc.virsat.fdir.core.matrix.iterator.LinearProgramIterator;
 import de.dlr.sc.virsat.fdir.core.matrix.iterator.MarkovAutomatonValueIterator;
 import de.dlr.sc.virsat.fdir.core.matrix.iterator.SSAIterator;
-import de.dlr.sc.virsat.fdir.core.metrics.FailLabelProvider.FailLabel;
 
 /**
  * Metric representing the covergent steady state probability
  * @author yoge_re
  *
  */
-public class SteadyStateAvailability implements IQuantitativeMetric, IDerivedMetric, IBaseMetric {
+public class SteadyStateAvailability implements IQuantitativeMetric, IBaseMetric {
 	public static final SteadyStateAvailability SSA = new SteadyStateAvailability();
 
 	/**
@@ -43,16 +39,6 @@ public class SteadyStateAvailability implements IQuantitativeMetric, IDerivedMet
 	@Override
 	public void accept(IBaseMetricVisitor visitor, SubMonitor subMonitor) {
 		visitor.visit(this);
-	}
-	
-	@Override
-	public void accept(IDerivedMetricVisitor visitor) {
-		visitor.visit(this);
-	}
-	
-	@Override
-	public Map<FailLabelProvider, Set<IMetric>> getDerivedFrom() {
-		return Collections.singletonMap(new FailLabelProvider(FailLabel.FAILED), Collections.singleton(Availability.INF_AVAILABILITY));
 	}
 	
 	/**
