+++ de.dlr.sc.virsat.model.extension.fdir/META-INF/MANIFEST.MF
@@ -12,7 +12,6 @@ Require-Bundle: org.eclipse.core.runtime,
  de.dlr.sc.virsat.fdir.core,
  de.dlr.sc.virsat.model.calculation,
  de.dlr.sc.virsat.fdir.storm,
- de.dlr.sc.virsat.external.lib.commons.math,
  de.dlr.sc.virsat.model.ext.core,
  de.dlr.sc.virsat.model.edit;visibility:=reexport
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
