+++ de.dlr.sc.virsat.model.extension.fdir/src/de/dlr/sc/virsat/model/extension/fdir/evaluator/DFTMetricsComposer.java
@@ -17,29 +17,17 @@ import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import org.apache.commons.math3.analysis.UnivariateFunction;
-import org.apache.commons.math3.analysis.integration.SimpsonIntegrator;
-import org.apache.commons.math3.analysis.integration.UnivariateIntegrator;
-import org.apache.commons.math3.analysis.interpolation.SplineInterpolator;
-import org.apache.commons.math3.analysis.interpolation.UnivariateInterpolator;
 import org.eclipse.core.runtime.SubMonitor;
 
 import de.dlr.sc.virsat.fdir.core.markov.modelchecker.ModelCheckingResult;
 import de.dlr.sc.virsat.fdir.core.metrics.Availability;
-import de.dlr.sc.virsat.fdir.core.metrics.Detectability;
-import de.dlr.sc.virsat.fdir.core.metrics.FailLabelProvider;
-import de.dlr.sc.virsat.fdir.core.metrics.FailLabelProvider.FailLabel;
 import de.dlr.sc.virsat.fdir.core.metrics.IBaseMetric;
 import de.dlr.sc.virsat.fdir.core.metrics.IBaseMetricVisitor;
-import de.dlr.sc.virsat.fdir.core.metrics.IDerivedMetric;
-import de.dlr.sc.virsat.fdir.core.metrics.IDerivedMetricVisitor;
 import de.dlr.sc.virsat.fdir.core.metrics.IQuantitativeMetric;
 import de.dlr.sc.virsat.fdir.core.metrics.MeanTimeToFailure;
-import de.dlr.sc.virsat.fdir.core.metrics.MeanTimeToDetection;
 import de.dlr.sc.virsat.fdir.core.metrics.MinimumCutSet;
 import de.dlr.sc.virsat.fdir.core.metrics.Reliability;
 import de.dlr.sc.virsat.fdir.core.metrics.SteadyStateAvailability;
-import de.dlr.sc.virsat.fdir.core.metrics.SteadyStateDetectability;
 import de.dlr.sc.virsat.model.extension.fdir.model.Fault;
 import de.dlr.sc.virsat.model.extension.fdir.model.FaultTreeNode;
 import de.dlr.sc.virsat.model.extension.fdir.model.VOTE;
@@ -51,13 +39,11 @@ import de.dlr.sc.virsat.model.extension.fdir.modularizer.Module;
  *
  */
 
-public class DFTMetricsComposer implements IBaseMetricVisitor, IDerivedMetricVisitor {
+public class DFTMetricsComposer implements IBaseMetricVisitor {
 	
 	private ModelCheckingResult composedResult;
-	private Map<FailLabelProvider, ModelCheckingResult> baseResults;
 	private long k;
 	private List<ModelCheckingResult> subModuleResults;
-	private double delta;
 	
 	/**
 	 * Compose operation for composable metrics. Composes the results of the sub modules.
@@ -74,27 +60,6 @@ public class DFTMetricsComposer implements IBaseMetricVisitor, IDerivedMetricVis
 		
 		for (IBaseMetric metric : metrics) {
 			metric.accept(this, subMonitor);
-		}
-		
-		return composedResult;
-	}
-	
-	
-	/**
-	 * Derivation operation of deriavable metrics. Calculates the metrics from
-	 * already calculated metrics in the given result
-	 * @param baseResults the given results
-	 * @param metrics the metrics to derive
-	 * @param delta the model checking delta
-	 * @return the derived results
-	 */
-	public ModelCheckingResult derive(Map<FailLabelProvider, ModelCheckingResult> baseResults, double delta, IDerivedMetric... metrics) {
-		this.composedResult = baseResults.get(new FailLabelProvider(FailLabel.FAILED));
-		this.baseResults = baseResults;
-		this.delta = delta;
-		
-		for (IDerivedMetric metric : metrics) {
-			metric.accept(this);
 		}
 		
 		return composedResult;
@@ -125,29 +90,7 @@ public class DFTMetricsComposer implements IBaseMetricVisitor, IDerivedMetricVis
 
 	@Override
 	public void visit(MeanTimeToFailure mttfMetric) {
-		int countFailRates = composedResult.getFailRates().size();
-		
-		if (countFailRates == 1) {
-			composedResult.setMeanTimeToFailure(Double.POSITIVE_INFINITY);
-			return;
-		}
-		
-		double[] x = new double[countFailRates];
-		for (int i = 0; i < countFailRates; ++i) {
-			x[i] = i;
-		}
-		
-		double[] y = new double[countFailRates];
-		for (int i = 0; i < countFailRates; ++i) {
-			y[i] = 1 - composedResult.getFailRates().get(i);
-		}
-		
-		UnivariateInterpolator interpolator = new SplineInterpolator();
-		UnivariateFunction function = interpolator.interpolate(x, y);
-		
-		UnivariateIntegrator integrator = new SimpsonIntegrator();
-		double integral = integrator.integrate(SimpsonIntegrator.DEFAULT_MAX_ITERATIONS_COUNT, function, 0, countFailRates - 1);
-		composedResult.setMeanTimeToFailure(integral * delta);
+		throw new UnsupportedOperationException("MTTF is not composable!");
 	}
 
 	@Override
@@ -185,33 +128,7 @@ public class DFTMetricsComposer implements IBaseMetricVisitor, IDerivedMetricVis
 			
 			Set<Set<Object>> productMinCutSets = minimumCutSet.cartesianComposition(allMinCuts);
 			composedResult.getMinCutSets().addAll(productMinCutSets);
-		}	
-	}
-
-
-	@Override
-	public void visit(Detectability detectabilityMetrc) {
-		ModelCheckingResult resultUnobservedFailure = baseResults.get(new FailLabelProvider(FailLabel.FAILED));
-		ModelCheckingResult resultObservedFailure = baseResults.get(new FailLabelProvider(FailLabel.OBSERVED));
-		detectabilityMetrc.derive(resultUnobservedFailure.getAvailability(), resultObservedFailure.getAvailability(), composedResult.getDetectabiity());
-	}
-
-
-	@Override
-	public void visit(MeanTimeToDetection meanTimeToDetectionMetric) {
-		ModelCheckingResult resultUnobservedFailure = baseResults.get(new FailLabelProvider(FailLabel.FAILED));
-		ModelCheckingResult resultObservedFailure = baseResults.get(new FailLabelProvider(FailLabel.OBSERVED));
-		double derivedMTTD = meanTimeToDetectionMetric.derive(resultUnobservedFailure.getMeanTimeToFailure(), resultObservedFailure.getMeanTimeToFailure());
-		composedResult.setMeanTimeToDetection(derivedMTTD);
-	}
-
-
-	@Override
-	public void visit(SteadyStateDetectability steadyStateDetectability) {
-		ModelCheckingResult resultUnobservedFailure = baseResults.get(new FailLabelProvider(FailLabel.FAILED));
-		ModelCheckingResult resultObservedFailure = baseResults.get(new FailLabelProvider(FailLabel.OBSERVED));
-		double derivedSteadyStateDetectability = steadyStateDetectability.derive(resultUnobservedFailure.getSteadyStateAvailability(), resultObservedFailure.getSteadyStateAvailability());
-		composedResult.setSteadyStateDetectability(derivedSteadyStateDetectability);
+		}
 	}
 	
 	/**
